requirements.txt
Flask
Flask-PyMongo
Flask-JWT-Extended
pymongo
python-dotenv
